// workshop/tools/ui/cf-app/app/src/api2/api.ts
const BASE_PATH = "http://localhost".replace(/\/+$/, "");

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
    csv: ",",
    ssv: " ",
    tsv: "\t",
    pipes: "|",
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
    (url: string, init?: any): Promise<Response>;
}

/**
 *
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
    url: string;
    options: any;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
    protected configuration: Configuration;

    constructor(configuration?: Configuration, protected basePath: string = BASE_PATH, protected fetch: FetchAPI = portableFetch) {
        if (configuration) {
            this.configuration = configuration;
            this.basePath = configuration.basePath || this.basePath;
        }
    }
};
// ...
export class DeviceDto extends EventEmitter {
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    id?: number;
    /**
     *
     * @type {boolean}
     * @memberof DeviceDto
     */
    isActive?: boolean;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    name?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    fields?: string;
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    user?: number;
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    project?: number;
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    flags?: number;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    state?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    host?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    port?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    protocol?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    logging?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    retry?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    blocks?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    commandSettings?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    settings?: string;
}
// ...
export interface InDeviceDto {
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    id?: number;
    /**
     *
     * @type {boolean}
     * @memberof InDeviceDto
     */
    isActive?: boolean;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    name?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    fields?: string;
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    user?: number;
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    project?: number;
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    flags?: number;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    state?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    host?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    port?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    protocol?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    logging?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    retry?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    blocks?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    commandSettings?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    settings?: string;
}
// ...
export class DevicesApi extends BaseAPI {
    /**
     *
     * @param {} [project] Group id for filter data by group. (default: empty)
     * @param {} [group] Group id for filter data by group. (default: empty)
     * @param {} [q] Text for search (default: empty)
     * @param {} [perPage] Number of results to return per page. (default: 10)
     * @param {} [curPage] A page number within the paginated result set. (default: 1)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DevicesApi
     */
    public apiDevicesGet(project?: number, group?: number, q?: string, perPage?: number, curPage?: number, options?: any) {
        return DevicesApiFp(this.configuration).apiDevicesGet(project, group, q, perPage, curPage, options)(this.fetch, this.basePath);
    }

    /**
     *
     * @param {} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DevicesApi
     */
    public apiDevicesIdDelete(id: number, options?: any) {
        return DevicesApiFp(this.configuration).apiDevicesIdDelete(id, options)(this.fetch, this.basePath);
    }

    /**
     *
     * @param {} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DevicesApi
     */
    public apiDevicesIdGet(id: number, options?: any) {
        return DevicesApiFp(this.configuration).apiDevicesIdGet(id, options)(this.fetch, this.basePath);
    }

    /**
     *
     * @param {} inDeviceDto
     * @param {} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DevicesApi
     */
    public apiDevicesIdPut(inDeviceDto: InDeviceDto, id: number, options?: any) {
        return DevicesApiFp(this.configuration).apiDevicesIdPut(inDeviceDto, id, options)(this.fetch, this.basePath);
    }

    /**
     *
     * @param {} inDeviceDto
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DevicesApi
     */
    public apiDevicesPost(inDeviceDto: InDeviceDto, options?: any) {
        return DevicesApiFp(this.configuration).apiDevicesPost(inDeviceDto, options)(this.fetch, this.basePath);
    }

    /**
     *
     * @param {} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DevicesApi
     */
    public apiDevicesStartIdPost(id: number, options?: any) {
        return DevicesApiFp(this.configuration).apiDevicesStartIdPost(id, options)(this.fetch, this.basePath);
    }

    /**
     *
     * @param {} id
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DevicesApi
     */
    public apiDevicesStopIdPost(id: number, options?: any) {
        return DevicesApiFp(this.configuration).apiDevicesStopIdPost(id, options)(this.fetch, this.basePath);
    }

}